src/category/domain/entities.py
from dataclasses import dataclass, field
from datetime import datetime
from typing import Optional

from src.__seedwork.domain.entity import Entity


@dataclass(kw_only=True, frozen=True, slots=True)
class Category(Entity):
    name: str
    description: Optional[str] = None
    is_active: Optional[bool] = True
    created_at: Optional[datetime] = field(default_factory=lambda: datetime.now())

    def update(self, name, description):
        self._set('name', name)
        self._set('description', description)

    def activate(self):
        self._set('is_active', True)

    def deactivate(self):
        self._set('is_active', False)
